+++ src/meta/src/hummock/compaction/picker/trivial_move_compaction_picker.rs
@@ -85,12 +85,12 @@ impl TrivialMovePicker {
                 input_levels: vec![
                     InputLevel {
                         level_idx: self.level as u32,
-                        level_type: LevelType::Nonoverlapping as i32,
+                        level_type: LevelType::Nonoverlapping,
                         table_infos: vec![trivial_move_sst],
                     },
                     InputLevel {
                         level_idx: self.target_level as u32,
-                        level_type: LevelType::Nonoverlapping as i32,
+                        level_type: LevelType::Nonoverlapping,
                         table_infos: vec![],
                     },
                 ],
