+++ src/meta/src/hummock/compaction/picker/ttl_reclaim_compaction_picker.rs
@@ -245,7 +245,7 @@ mod test {
             ),
             Level {
                 level_idx: 4,
-                level_type: LevelType::Nonoverlapping as i32,
+                level_type: LevelType::Nonoverlapping,
                 table_infos: vec![
                     generate_table_with_ids_and_epochs(2, 1, 0, 100, 1, vec![2], expired_epoch, 0),
                     generate_table_with_ids_and_epochs(
