+++ src/meta/src/hummock/compaction/selector/mod.rs
@@ -127,7 +127,7 @@ pub mod tests {
         levels.l0.as_mut().unwrap().total_file_size += sst.file_size;
         levels.l0.as_mut().unwrap().sub_levels.push(Level {
             level_idx: 0,
-            level_type: LevelType::Overlapping as i32,
+            level_type: LevelType::Overlapping,
             total_file_size: sst.file_size,
             uncompressed_file_size: sst.uncompressed_file_size,
             sub_level_id: sst.get_sst_id(),
@@ -145,7 +145,7 @@ pub mod tests {
             .sub_levels
             .last_mut()
             .unwrap()
-            .level_type = LevelType::Nonoverlapping as i32;
+            .level_type = LevelType::Nonoverlapping;
     }
 
     pub fn push_tables_level0_nonoverlapping(levels: &mut Levels, table_infos: Vec<SstableInfo>) {
@@ -158,7 +158,7 @@ pub mod tests {
         levels.l0.as_mut().unwrap().total_file_size += total_file_size;
         levels.l0.as_mut().unwrap().sub_levels.push(Level {
             level_idx: 0,
-            level_type: LevelType::Nonoverlapping as i32,
+            level_type: LevelType::Nonoverlapping,
             total_file_size,
             sub_level_id,
             table_infos,
@@ -244,7 +244,7 @@ pub mod tests {
             .sum();
         Level {
             level_idx,
-            level_type: LevelType::Nonoverlapping as i32,
+            level_type: LevelType::Nonoverlapping,
             table_infos,
             total_file_size,
             sub_level_id: 0,
@@ -267,7 +267,7 @@ pub mod tests {
                 .enumerate()
                 .map(|(idx, table)| Level {
                     level_idx: 0,
-                    level_type: LevelType::Nonoverlapping as i32,
+                    level_type: LevelType::Nonoverlapping,
                     total_file_size: table.file_size,
                     uncompressed_file_size: table.uncompressed_file_size,
                     sub_level_id: idx as u64,
@@ -289,7 +289,7 @@ pub mod tests {
                 .enumerate()
                 .map(|(idx, table)| Level {
                     level_idx: 0,
-                    level_type: LevelType::Nonoverlapping as i32,
+                    level_type: LevelType::Nonoverlapping,
                     total_file_size: table.iter().map(|table| table.file_size).sum::<u64>(),
                     uncompressed_file_size: table
                         .iter()
@@ -324,7 +324,7 @@ pub mod tests {
                 .enumerate()
                 .map(|(idx, table)| Level {
                     level_idx: 0,
-                    level_type: LevelType::Overlapping as i32,
+                    level_type: LevelType::Overlapping,
                     total_file_size: table.iter().map(|table| table.file_size).sum::<u64>(),
                     sub_level_id: idx as u64,
                     table_infos: table.clone(),
